hmi2mid.c, hmp2mid.c (READ_INT32LE): avoid signed integer overflow.

+++ src/hmi2mid.c
@@ -505,7 +505,7 @@ static uint32_t ConvertListToMTrk(struct hmi_ctx *ctx, midi_event *mlist) {
     return (i);
 }
 
-#define READ_INT32LE(b) ((uint32_t)((b)[0] | ((b)[1] << 8) | ((b)[2] << 16) | ((b)[3] << 24)))
+#define READ_INT32LE(b) ((uint32_t)(b)[0] | ((uint32_t)(b)[1] << 8) | ((uint32_t)(b)[2] << 16) | ((uint32_t)(b)[3] << 24))
 
 int _WM_hmi2midi(const uint8_t *in, uint32_t insize,
                  uint8_t **out, uint32_t *outsize) {
